requirements.txt
# Core ML & Data
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
shap>=0.43.0
imbalanced-learn>=0.12.0
joblib>=1.3.0

# LangChain & LangGraph
langchain>=0.3.0
langchain-openai>=0.2.0
langgraph>=0.2.0
langsmith>=0.1.0

# Visualization
matplotlib>=3.7.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
kagglehub>=0.3.0

# Web UI
streamlit>=1.28.0
